requirements.txt
# Vision-to-Action Dependencies

# Core deep learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Computer vision utilities
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0

# imgshape (local symlink - editable install)
-e ./imgshape

# Data handling
pydantic>=2.0.0
PyYAML>=6.0
pandas>=2.0.0

# LLM / Cognition
google-generativeai>=0.3.0

# API / Networking
requests>=2.31.0
aiohttp>=3.9.0

# Storage
# Optional: database connectors if needed
# psycopg2-binary>=2.9.0  # PostgreSQL
# pymongo>=4.0.0  # MongoDB

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter (optional for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
